完整版本提交

+++ xxl-code-generator-admin/src/main/resources/templates/xxl-code-generator/service_impl.ftl
@@ -18,42 +18,95 @@ public class ${classInfo.className}ServiceImpl implements ${classInfo.className}
 	private ${classInfo.className}Dao ${classInfo.className?uncap_first}Dao;
 
 	/**
+    * 新增或修改
+    */
+	@Override
+	public ReturnModel insertOrUpdate(${classInfo.className} ${classInfo.className?uncap_first}) {
+	  ReturnModel result = new ReturnModel();
+	  if (${classInfo.className?uncap_first} == null) {
+	     return result;
+	  }
+	  if(null != ${classInfo.className?uncap_first}.getId() && !"".equals(${classInfo.className?uncap_first}.getId())){
+		//修改
+       result =  update(${classInfo.className?uncap_first});
+	  }else{ //新增
+       result = insert(${classInfo.className?uncap_first});
+	  }
+	  return result;
+	}
+
+	/**
     * 新增
     */
 	@Override
-	public ReturnT<String> insert(${classInfo.className} ${classInfo.className?uncap_first}) {
+	public ReturnModel insert(${classInfo.className} ${classInfo.className?uncap_first}) {
+      ReturnModel result = new ReturnModel();
 	  if (${classInfo.className?uncap_first} == null) {
-	    return new ReturnT<String>(ReturnT.FAIL_CODE, "必要参数缺失");
+	     result.addDefaultModel("404","必要参数缺失");
+		 return result;
 	  }
 	  ${classInfo.className?uncap_first}Dao.insert(${classInfo.className?uncap_first});
-	  return ReturnT.SUCCESS;
+	  result.setSuccess(true);
+	  return result;
 	}
 
 	/**
 	* 删除
 	*/
 	@Override
-	public ReturnT<String> delete(int id) {
+	public ReturnModel delete(String id) {
+      ReturnModel result = new ReturnModel();
+	  if(null == id || "".equals(id)){
+ 	 	result.setMsg("id不能为空！");
+		return result;
+	  }
 	  int ret = ${classInfo.className?uncap_first}Dao.delete(id);
-	  return ret>0?ReturnT.SUCCESS:ReturnT.FAIL;
+	  if(ret > 0){
+		result.setSuccess(true);
+		return  result;
+	  }else{
+		return  result;
+	  }
 	}
 
 	/**
-	* 更新
+	* 修改
 	*/
 	@Override
-	public ReturnT<String> update(${classInfo.className} ${classInfo.className?uncap_first}) {
+	public ReturnModel update(${classInfo.className} ${classInfo.className?uncap_first}) {
+      ReturnModel result = new ReturnModel();
 	  int ret = ${classInfo.className?uncap_first}Dao.update(${classInfo.className?uncap_first});
-	  return ret>0?ReturnT.SUCCESS:ReturnT.FAIL;
+	  if(ret > 0){
+		result.setSuccess(true);
+		return  result;
+	  }else{
+		return  result;
+	  }
 	}
 
 	/**
 	* 根据Id查询
 	*/
 	@Override
-	public ${classInfo.className} load(int id) {
-	  return ${classInfo.className?uncap_first}Dao.load(id);
+	public ReturnModel load(String id) {
+ 	  ReturnModel result = new ReturnModel();
+ 	  if(null == id || "".equals(id)){
+ 	 	result.setMsg("id不能为空！");
+		return result;
+	  }
+	  ${classInfo.className} ${classInfo.className?uncap_first} = ${classInfo.className?uncap_first}Dao.load(id);
+ 	  result.addDefaultModel("value",${classInfo.className?uncap_first});
+	  result.setSuccess(true);
+	  return result;
 	}
+
+ 	/**
+    * 全部查询
+    */
+    public List<${classInfo.className}> getAll(){
+		List<${classInfo.className}> list = ${classInfo.className?uncap_first}Dao.getAll();
+		return list;
+	};
 
 	/**
 	* 分页查询
